fix(Collisions): only track active containing transform collisions
The ActiveCollisionsContainer was tracking every collision between
nested colliders due to changes in how the CollisionNotifier worked
but this caused issues as the active collisions should only be based
on the overall containing colliding Transform otherwise unwanted
artifacts are produced such as untouch events when moving between
colliders even though the object is still fundamentally being touched.

This then had a compound affect on publishing to all of the nested
colliders which would cause performance issues.

The fix is to track each individual collision that is received from
the CollisionNotifier but to keep a dictionary of which collisions are
associated to the overall containing Transform. This way the
ActiveCollisionsContainer logic and events can be limited to only
happen when a completely new collision on another containing Transform.

This brings the functionality back in line with how it worked before
the CollisionNotifier changes, whilst honouring the changes to the
ability to allow a Collider to be disabled and to tell the
CollisionTracker that the collision has ended.