feat(Collision): clear active collisions from publisher
The ActiveCollisionPublisher can now have any stored active collisions
cleared via a public method. This is in addition to the ClearPayload
method which also clears the SourceContainer which isn't always ideal.

The iteration of the Payload.ActiveCollisions is now stored in a
holding list in case the ClearActiveCollisions is called during the
enumeration process.