feat(Type): add collapsible UnityEvent drawer
A new custom PropertyDrawer that overrides the UnityEventDrawer has
been added to allow the ability to wrap a UnityEvent inspector control
in a foldout so the UnityEvent control can be collapsed.

Reflection is required to override the default UnityEventDrawer
inspector drawing ue to the way Unity creates the UnityEventDrawer
internally.