fix(Data): display unsupported types warning on Observable List drawer
The Observable List Property Drawer only supports inline elements of
type UnityEngine.Object so any other type that does not derive from
this will throw an exception. The fix is to simply check the inherited
type and to display a warning box if it doesn't match the supported
type.

The Property Height has also been override to remove unwanted leading
vertical space from the property drawer.