fix(Data): utilize SerializedObject data to draw list elements
There is an issue with manually building the list out of the
instantiated list object data as it will never pass through Unity's
serializer and therefore any changes will not be serialized to disk
meaning the control doesn't actually persist changes when the
scene is saved.

This fix removes all of the custom code and simply gets a
SerializedObject of the referenced property and uses the Unity
built in PropertyField to render the field which will ensure the
property changes are serialized to disk.

This does have the drawback that the rendered property drawer does
not utilize the custom ObservableListEditor which handles manually
firing the Observable List events when changes occur. This means
any changes to the Elements list in a reference property (i.e. using
this custom property drawer) will not emit any of the change events
found on the Observable List component. To counter this issue, the
shortcut inspector properties are disabled at runtime to prevent
any confusion about misuse.

Whilst this isn't an ideal solution, it isn't a breaking issue as
the underlying Observable List component still will emit the events
when changed in the inspector so any such requirements would just
simply need the actual Observable List component to be worked on
at runtime in the editor and this Property Drawer is simply considered
a nice shortcut helper for setting up references quickly in the editor.