feat(Visual): add events for transitioned state of overlay
The CameraColorOverlay now has additional events and state for whether
the overlay is currently in an Add or Remove transition and it emits
an event when it reaches the end of either the Add or Remove transition
state with the existing Added/Removed events being there to signal when
the transition process was initiated.

The Added/Removed event names aren't ideal but changing them will
introduce an unwanted breaking change which doesn't seem worth it
right now.