fix(MomentProcessor): allow PreRender to work with Unity SRP
The new Unity Scriptable Render Pipeline introduces a new way of
handling the event for when a camera begins rendering, meaning the
existing `Camera.onPreRender` event is no longer utilized. This
causes the `PreRender` moment to fail in the MomentProcessor because
the camera no longer has an `onPreRender` event when using the
scriptable render pipeline. This has been replaced by using the
`RenderPipelineManager` and the `beginCameraRendering` event which
can be used in place of the `onPreRender` if an SRP is being used.

This fix simply checks to see if the graphic settings are using a
scriptable render pipeline and uses the appropriate event for the
PreRender moment.