fix(Process): restrict SRP support to 2019.1 and above
A compile error was occurring because Unity 2018 did not support the
SRP related calls outside of an experimental namespace. It's not a good
idea to support experimental Unity solutions so the fix for this
compile error is to just only support SRP on Unity 2019.1 and above as
the required functionality is standard in 2019.1 and above.

A warning will be logged if SRP is used on 2018.