fix(Tracking): remove ancestor position when calculating rotation
The TransformPositionDifferenceRotation had an issue where if the
source and target GameObjects were descendants of another GameObject
that was also moving, then the source position and previous position
would be further away due to the move in their ancestor GameObject
causing the rotation calculation to be incorrect. This fix provides
the ability to inject the ancestor GameObject to ensure the calculation
is still correct.