feat(Velocity): multiply velocity by vector rather than float
*Breaking Change*

The VelocityMultiplier only allowed for a float to be multiplied
to the whole velocity vector making the multiplication very linear
and not able to affect different parts of the velocity by different
values.

This has now been changed so the multiplier is actually a Vector3,
meaning it can affect different elements with different multipliers.

This is also useful if a certain velocity axis wants to be removed
by simply multiplying that axis by 0.