ci: implement continuous delivery
Since Zinnia is now a UPM package it should automatically be
released as one. This change adds automatic creation of UPM packages
for Zinnia, including automatic SemVer-styled versioning based on
the commit messages. A release is both uploaded to the ExtendReality
npm GitHub registry as well as an GitHub release (archived .zip).