feat(Extension): determine if Behaviour is in a valid active state
The IsValidState extension method can be used to determine if a
Behaviour is in the appropriate valid active state depending on
whether the component is enabled and/or whether the GameObject the
component resides on is active within itself or within the scene
hierarchy.

This mechanism is a copy of the Malimbe `[RequiresBehaviourState]` tag
but can be used inline in methods rather than just a blanket early
return from the entire method.