feat(Collection): allow query operations on list to run when inactive
The query operations on the ObservableList such as Contains and
IsListEmpty were previously encapsulated by the
`[RequiresBehaviourState]` tag meaning a list contents check could
only be performed if the list component was active in the scene.

This blanket denial of operation is too heavy handed as just because
a list object is inactive it shouldn't mean a contains check returns
false because the list, even though inactive, still can contain the
object to check for.

It is still correct that components shouldn't perform an action when
they are disabled, so if a list is queried then it should return
the correct result but shouldn't perform any action, such as emitting
any appropriate events.

This solution uses the new Behaviour extension `IsValidState` to
replicate the `[RequiresBehaviourState]` functionality within the
method body to allow the correct return value but to prevent the
event actions from occurring.