feat(Rule): determine if the rule component state can auto reject
A new base abstract Rule class has been added that all existing
MonoBehaviour rules now extend from. And on this new base Rule is
the concept of being able to set whether the rule can automatically
reject any request based on the state of the Rule component.

Previously, if a Rule component was disabled or on an inactive
GameObject then it would always reject the `Accepts` request, but this
isn't always the requirement when using a Rule. The new
`AutoRejectStates` enum flag allows a Rule to specify what states the
rule can be in to automatically reject any request.

All of the tests have been updated to show this new state selection
option in practice.