feat(Haptics): option to allow all haptic processes in list to process
The HapticProcessor now has a `CeaseAfterFirstSourceProcessed` option
which is defaulted to true to ensure existing default behaviour does
not change, which is when the first active Haptic Process is processed
then the HapticProcessor would cease to process any further Haptic
Processes.

With this new option, this default behaviour can be turned off and
all Haptic Processes listed in the HapticProcessor can be processed
if they are active in the scene.