fix(README.md): provide more concise release data and update info
The Releases section has been removed and is now just a simple badge
at the top of the README. There has been an additional section in
`Getting Started` on how to update the package via the Unity Package
Manager.

The links have also been ordered in the order of appearance in the
document.