fix(Collision): always add/remove the CollisionTrackerDisabledObserver
The CollisionTrackerDisabledObserver is now always added and removed
when a collision occurs rather than before where it was only added or
removed if that collision state was being listened to.

The previous way would cause the CollisionTrackerDisabledObserver to
either not be added at all or be added multiple times because it was
never being removed.