feat(ObjectPointer): option to disable destination on no collision
The ObjectPointer can now disable the destination pointer element if
there is no collision data from the RayCast.

Also, the ObjectPointer test was incorrect as the pointer elements
were set up incorrectly where the container was assigned to the mesh
and vice versa. This has now been corrected and the tests updated
to reflect this change.