fix(Data): ensure ObservableListEditor raises events with inspector
There was an issue where the ObservableListEditor was not raising
events when the Elements property was changed in the inspector
due to an issue with the underlying Malimbe custom inspector would
only call the `BeforeChange()` and `AfterChange()` methods when
the component had a ChangeHandler attribute associated with it.

The ObservableListInspector requires the `BeforeChange()` and
`AfterChange()` methods to be executed to raise the appropriate
events, the correct way to do this is to override the
`FindChangeHandlersMethods()` method and ensure that the
ObservableList Elements property is considered something that
is a ChangeHandler.