feat(Rule): add rule to check if a string pattern is in a string list
The StringInListRule allows a string pattern to be specified to match
against any string found in an associated String Observable List that
is a component on the given GameObject.

This can replace the need to use the AnyComponentTypeRule and creating
dummy scripts for tags as the string list can be used as the tag name
that is then picked up by the rule.