feat(Action): add method for action to receive its own default value
The new `ReceiveDefaultValue` method is a shortcut for the `Receive`
method but it simply makes the Action receive its own default value.

This can be used to programmatically get an Action to call its own
deactivation event if the Action is already activated without needing
to know the type of the concrete action. This is useful when dealing
with Actions in their generic abstract form but wanting them to
emit their deactivated state for whatever reason, such as a linked
GameObject becoming disabled.