fix(Extraction): ensure extraction cannot be mutated if disabled
The main `Extract()` method should do a check to see if the component
is active and enabled and if its not then it should force set the
`Result` to `null` across all Extractors.

Also, the `RequiresBehaviourState` attribute has been added to the
`Extract` methods that allow the data to be passed in via a parameter
as they were allowing mutation to the source even when the component
was inactive.

Finally, any extractors that didn't have the `Extract` methods that
allowed a parameter have been updated to include these methods too.