feat(Extraction): consolidate all extractors into 1 base class
(Nearly) all of the extractors now all inherit from a single
ValueExtractor class which contains most of the logic ensuring
extractors all have a standard API.

There are a couple of extractors that have multiple events for
extraction and these don't fit into this model yet so have been
left out. These are:

* ObjectDistanceComparatorEventDataExtractor
* TransformPropertyApplierEventDataExtractor

These will be updated in the future to provide individual extrators
that can then follow the standard Extractor pattern.

All Extractors now also implement `iProcessable` so can all be used
with a `MomentProcessor`.

All Extractors also invoke a `Failed` event which is raised when the
extractor has failed to extract the value. This is to ensure the
`PlayAreaDimensionExtractor` still has the relevant events required
and can fit in the standard Extractor pattern. Plus, having a `Failed`
event is useful to know when an Extractor has failed.

The `TransformPropertyExtractor` has been renamed to
`TransformVector3PropertyExtractor` which originally existed within
the same file but now it can simply extend the `Vector3Extrator` but
the old extractor had a separate property for storing the result of
the extraction called `LastExtractedValue`. This property is still
available but it has been deprecated and the `Result` field should
now be used to get the extracted value.