fix(Extraction): flip extraction logic for local direction
The TransformDirectionExtractor had the `UseLocal` logic the wrong
way round. It was returning the global `Vector3.<direction>` if
`UseLocal` was true and returning the direction of the `Source` if
`UseLocal` was false.

This doesn't make sense because `Vector3.<direction>` _is_ the global
direction whereas the direction of the `Source` is technically local
to the `Source`.

This is also technically a change that can cause breaks because the
logic is now flipped. But rather than do a clever deprecation or
anything, it's probably just better to handle complaints as its just
fixed by checking (or unchecking) the `UseLocal` property.