feat(Extraction): add extractors for SurfaceData and RaycastHit
A collection of extractors that firstly extract the RaycastHit data
from a SurfaceData and then a collection of extractors that can
extract specific data from the RaycastHit.

This new collection means the existing
SurfaceDataCollisionPointExtractor becomes obsolete as it is too
specific for the extractor pattern and can be achieved by first
extracting the RaycastHit from the SurfaceData and then extracting
the RaycastHit.point from the RaycastHit output.