feat(Cache): provide cache operations for common data types
The Cache operation allows a data type value to be stored in a cache
and then an appropriate event is raised when the value is updated.

If the value is considered equal then the Unmodified event is
emitted. However, if the values are not equal then the Modified event
is emitted.