feat(Tracking): allow previous position to be valid located surface
The SurfaceLocator would always not consider a valid location if
the previous location was the same as the current location. This new
option allows that equality check to be ignored.

Also, the equality threshold is no longer a constant and instead a
value that can be changed via the public property.