fix(Extraction): override ClearSource method for generics
There is an issue with using the Malimbe generated ClearSource method
when using it with a generic type property as the actual type may
not be a nullable type and the MemberCleared weaved code always tries
to set it to `null`.

In this regard, it is better to simply override the method and set
the property to `default`. This is not done in Malimbe as providing
default as the weaved code is extremely tricky.