feat(Association): determine association by platform, sdk and model
The PlatformDeviceAssociation component allows the determining of the
hardware based on matching patterns of the current platform, the
SDK loaded by Unity and the model type.

As all of these are patterns, then they can be used to search for
anything or specific types of setup.

The LoadedXrDeviceAssociation has now been deprecated as it offers
a subset of this new functionality but isn't as powerful.