feat(Conversion): ability to convert between float and normalized float
The FloatToNormalizedFloat allows a float value to be converted into a
normalized float (between 0f and 1f) and the NormalizedFloatToFloat
allows for a normalized float to be converted to a float value based
within the range provided.

The FloatToBoolean has also had the min/max limit removed so any float
value can be used to check to see if it should be within the positive
bounds.