feat(Proxy): expose payload as public property
The Payload property on the Event Proxy has now been exposed as a
serialized public property so it can be set via the inspector or
via external code or Unity event. This payload can then be emitted
at a later stage by calling the existing `EmitPayload` method.

There is also a `ClearPayload` method for setting the Payload
property back to the `default` value.