feat(Tracking): apply rotation property based on angular velocity
The new RotateAroundAngularVelocity component will modify a target's
rotation property by rotating around the offset using the angular
velocity of the given VelocityTracker as the angles in which to rotate
per frame.

This can be used to simulate turning something with the rotation of
the controller, such as screwing in a screw.