feat(Observer): add observable property data types
The new ObservableProperty types allow a data type property to be
set in the inspector but any change to that property is observed
and an appropriate event is raised for the state of the property.

This is a better implementation of the ValueCache components which
didn't really offer much other than storing a value via code and
raising events when that value was modified. The ValueCache
components have now been deprecated with the new ObservableProperty
counterparts being a much better choice.