feat(Mutation): allow limiting axes on mutator offsets
The TransformPositionMutator has a Facing Direction offset that can
be provided which is used to offset the forward direction based on
the rotation of this optional offset GameObject. It is now possible
to determine which axes of that offset should be used when utilizing
the rotation data as there are occurrences where not all of the offset
rotations are wanted.

The TransformEulerRotationMutator has an Origin offset that can be used
to rotate around a different pivot point position. It is now possible
to determine which axes of direction should be used from the optional
pivot point as sometimes not all axes may be required.