feat(Extension): add new data type extensions for common calculations
A common calculation is finding a fine grain distance between two
points (either Vector2 or Vector3) where the tolerance is also given
in the same type as opposed to just a simple `float`.

The new Vector2.WithinDistance and Vector3.WithinDistance offer this
via the relevant extensions.

Another common calculation is converting euler degrees to signed
degrees, such as 270' is actually equivalent to -90'. This helps when
doing greater than or less than comparisons as a negative rotation of
-90' is less than a rotation of 0' whereas 270' as a number would be
greater than 0'.

The `Vector3.UnsignedEulerToSignedEuler` will convert the current
Vector3 of euler angles into a Signed Euler (-180' to 180f) using the
new `float.GetSignedDegree` which simply converts a Euler angle into
the -180' to 180' range.

These are then used to provide new Transform extensions for:

* `Transform.SignedEulerAngles`
* `Transform.SignedLocalEulerAngles`

Which simply return the respective Euler or Local Euler angle for the
Transform but in this signed format.