feat(Yield): add ability to emit an event after a yield instruction
The new Yield events provide the ability to trigger some action after
a yield instruction has completed such as seconds passed or at end of
frame.

This can be used in conjunction with the Proxy events to first store
the payload in the Proxy then trigger the emit after the yield
instruction has completed.