feat(structure): provide mechanism to change properties via UnityEvents
Some property types cannot be changed via UnityEvents as they are not
supported in the UnityEvent inspector, such as Enums, Vectors and
Vector3State.

This has been fixed by adding custom setter methods that can be called
via the UnityEvent inspector using primitive types that are supported
to allow this data to still be set.