fix(Attribute): record MinMaxRange value changes on prefab instance
There is an issue where the MinMaxRange control will reset the value
back to the previous value when it is used within a prefab.

The solution seems to be to record the prefab instance property
modification after the custom FloatRange value has been set through
the Supyrb `SetValue` extension, which doesn't set the value via
the SerializedProperty because that is not supported in Unity on custom
data types.

The issue only seems to present itself when changing the value between
varying negative values:

* -0.5
* -0.2
* -0.5 (reverts to 0)