feat(Tracking): add container for successful published consumers
The ActiveCollisionRegisteredConsumerContainer is linked to a
ActiveCollisionPublisher and its job is to store a collection
of consumers that the publisher has successfully published to.

This means the publisher still doesn't know directly who it is
publishing to, but holds an audit log of which consumers have
successfully received the payload in the past.

The ActiveCollisionConsumer now also holds a reference to the
top level container (i.e. the colliding rigidbody/collider GameObject)
that the component is residing under.