feat(Cast): provide ability to determine validity by target point
The new TargetPointValidity rule allows a type of Vector3Rule to be
used to determine if the PointsCast target point should be considered
as a valid target.

This can be used with the new NavMeshRule to determine if a target
point from the points cast is within the NavMesh boundaries.