fix(Extraction): extract correct direction based on UseLocal setting
The UseLocal setting was not returning data that made sense. The
`UseLocal = false` was just returning the world direction Vector which
had nothing to do with the `Source` so it made little sense to have it
part of the TransformDirectionExtractor.

The `UseLocal = true` setting was actually returning the world
direction of the `Source` so this is now the `UseLocal = false`
operation as `Source.transform.forward` is actually the world forward
direction of the `Source` even if it is a nested GameObject with
multiple parent rotations.

This means the `UseLocal = true` setting is now the local direction of
the `Source` GameObject so if it is a nested child with an additional
parent rotation then the local forward is purely the local rotation
of the GameObject multiplied by the world forward constant
(Vector3.forward).

Whilst this will introduce a breaking change, it's actually a fix for
operations that were not correct, so it's not a feature including
breaking changes, it's actually fixing inaccurate behaviour.