feat(Tracking): ability to determine surface validity by target point
The new TargetPointValidity rule allows a type of Vector3Rule to be
used to determine if the located point of the RayCast in the
SurfaceLocator should be considered a valid target.

This can be used with the NavMeshRule to determine if a point on a
surface is within the NavMesh boundaries.

The stub Vector3 rule has been moved out of the StraightLineCastTest
and into its own stub so it can be used in multiple places. This is
instead of testing with the actual NavMeshRule.