fix(Tracking): calculate angular velocity angle optionally in degrees
The RigidbodyAngularVelocity component was updated to calculate in
radians only. Whilst, this makes sense for the most part, it does not
create a good outcome in all situations. The component has been updated
so it can optionally use radians or degrees and it can also take in
to consideration the offset to determine the new centre of mass to
calculate around.

The divergence checking code was also faulty, so this has been fixed
as part of this commit.