feat(Utility): add custom editor to create multi part components
The ObservableListComponentGeneratorEditorWindow provides an easy way
of creating components that contain nested Observable Lists by using
a simple inspector window that provides a drop down list of available
components to create. Selecting the component and clicking the button
will add a GameObject to the scene with the desired component and then
it adds a nested GameObject that contains the relevant Observable list.

This makes adding components that rely on observable lists easier to
understand and quicker to achieve.