fix(Association): provide alternative mechanism for getting device name
Unity 2020.2 has removed the XRDevice.model property and in place it
attempts to use InputDevice.name. This fix attempts to get the
InputDevice for the XRNode.Head and uses the InputDevice.name for the
retrieved device.

These names returns are piped via the XR Management System and are
sometimes nonsensical values due to the vendors not providing correct
string names in their XR Plugin.