fix(Collection): prevent Added event being raised twice on start
There was an issue where if the ObservableList was added to in the
OnEnable of another script then the Added event be raised as the
ObservableList was enabled, but Start is called after OnEnable so
when the Start method was called it would then loop through the
list and raise the events for everything that existing in the list
the first time it was loaded.

This fix just stops the normal Add method from raising events if
Start hasn't been called.