feat(Data): overide Equals on data types
The base data types have had their Equals method overridden so they
perform a logical based equality check on the contents of the object
when the properties are the same. This allows these data types to be
logically compared such as in the use of Lists in where a TransformData
object with the same properties should be considered the same object.

The tests have been updated to test the equality of objects.