feat(Tracking): option to prevent collision notification on disable
A new option to the CollisionNotifier has been added that will prevent
collision tracking to occur if the component is disabled. Previously,
all collisions would continue to process even if the component was
disabled because the Unity OnCollision/Trigger[XXX] methods always run
even on a disabled component.

There are times where you may not want this to be the case so a new
property has been added that will prevent collision processing when
the component is disabled. It is still possible to have the component
disabled and collisions processed if this new property is set to true
(which is is by default so no noticable changes will occur).