feat(Rule): add new rule types for primitives
A collection of new rule types have been added that deal with primitive
object types such as floats, ints, strings and vectors.

These provide the basis for applying rules against primitive types such
as determining if a number is in a range, or whether a type is equal
to another type (within a given tolerance).