feat(Data): provide get method and obtained event on observable list
The ObservableList components now have a `Get(index)` method that
retrieves an element from the collection List and uses index clamping
so an invalid index will not throw an error, instead an index too high
will return the last element in the List and a negative index will
walk the list in reverse (e.g. -2 gets the second from last element
from the list).

There is also an accompanying `Obtained` event that emits the element
that was retrieved in the `Get` call. If `Get` is attempted on an
empty collection then it will result in an exception being thrown.