fix(Tracking): ignore extra trigger enter/exit on kinematic change
Unity 2019.3 introduced an undocumented change brought about by
PhysX 4.11 where when a rigidbody kinematic state is changed then
Unity will re-fire the OnTriggerExit and OnTriggerEnter events even
though the collision between colliders did not end. This causes
issues as extra events are raised that do not reflect what is actually
happening in the scene.

The fix involves being able to mark a rigidbody attached to a collider
as about to have its kinematic state changed, this is then stored and
upon the OnTriggerExit being called (in the same frame) it will ignore
the internal logic and then ignore the subsequent OnTriggerEnter logic
(which is also done in the same frame) and then cleans up the rigidbody
cached state so it knows that additional exits are valid.

This is only applied in Unity 2019.3 and above and can be disabled at
runtime if the mechanism of how Unity 2019.3 works is required.