fix(Tests): move some tests to run as UnityTest to call unity lifecycle
It would seem that a version of Unity has stopped calling Unity
lifecycle events such as Start in playmode tests under the normal
`[Test]` structure and now they require to be run as `[UnityTest]`
instead. It is not clear from the Unity changelogs when this change
was introduced, but it is definitely needed for Unity 2019.4.