feat(Tracking): allow local space rotation around angular velocity
When using an angular transform drive as a handleless door knob,
this class is used to track the interactor's "wrist" rotation,
but as it is now the this only works when drive/interactor are
aligned with world space.  By using this new option, the angular
velocity will be transformed to the target's space, and thus
applied properly.