feat(Tracking): improve direction modifier rotation
The DirectionModifier component has been improved so the rotation
of the Target now works if the Pivot is in front of the LookAt
GameObject and it is also possible to specify a TargetOffset for
handling non LookAt snapping rotations. It is also now possible
to determine which object to use for the world up in the LookRotation
process.

The `PreventLookAtZRotation` logic has been removed as it no longer
seems to be required after this new functionality has been added.