fix(Tracking): prevent kinematic state fix firing multiple times
The kinematic state change fix that was introduced to fix the PhysX
change in Unity 2019.3 was causing an issue with compound colliders
because the OnTriggerXXX will fire for every change in a compound
collider and not just at the containing Rigidbody.

This would mean that with compound colliders it would get out of
sync with the state changed check when the kinematic state changed.

This has been fixed now by checking if the given collider in the
OnTriggerExit method has already been fired (on the containing
Rigidbody) within the same frame, and if it has then this is most
likely a compound collider causing the event so they get ignored.