fix(Visual): prevent fade mesh getting destroyed on disable
The movement of the mesh overlay seems to cause the component to get
disabled and then re-enabled causing the mesh to not be available
at some points, which means the fade doesn't work.

Instead of destroying the mesh when the component is disabled, it now
just disables the renderer component and makes the destroy mesh method
public so it can be manually destroyed.