fix(Collection): remove custom list editor for Unity 2020.3 or newer
Unity have fixed a bug in Unity 2020.3.24f to do with reorderable lists
but that fix has caused the custom observable list editor to no longer
display any of the elements of the list and no longer understands if
the list is collapsible or not.

I can't be bothered to put the effort in to find out what Unity have
done or where this has broken, it's a constant struggle of chasing
after Unity changes so instead the feature of custom list editors
is now just disabled for Unity 2020.3 and above.

It can be added back in with a custom script define symbol of:

* `ZINNIA_USE_CUSTOM_LIST_EDITOR`

This will make the custom list editor appear again, however it will
no longer be a collapsible element.

The custom list editor can also now be completely disabled by using
the script define symbol of:

* `ZINNIA_IGNORE_CUSTOM_LIST_EDITOR`

This will completely disable the custom list editor functionality.