feat(Tracking): provide ability to delay collision stay processing
The CollisionTracker now has an option to set a delay interval for
the `Collision` and `Trigger` `Stay` methods to delay processing
the next stay state until the delay interval has passed.

This means it is possible to not have to process the stay state
every single fixed update frame as this may be too frequent for
certain circumstances and cause too much of a processing overhead.