feat(Association): define associations with generic patterns
The PlatformDeviceAssociation component has been deprecated due to the
limited ways it can form a device association via hard coded properties
related to the restricted supported types.

The new RuleAssociation component replaces it with a completely generic
approach that will determine activation status on the GameObject
collection if the given rule passes.

This then works with the new PatternMatcherRule which takes a
collection of new PatternMatcher components that will attempt to match
a give pattern string against a selected source property from a
given Unity namespace that contains useful data.

The existing associations in the PlatformDeviceAssociation exist now
as their own PatternMatcher components:

* ApplicationPatternMatcher
* XRSettingsPatternMatcher
* XRDevicePatternMatcher

These provide the same functionality as before, but also offer
additional properties that can be used to match against.

Thes PatternMatcher components can be set up with the
PatternMatcherRule and then that rule set up as the Rule in the
RuleAssociation to provide the same functionality but with much
more flexibility.

A new SystemInfoPatternMatcher is also included that can use many
property types from the `UnityEngine.SystemInfo` class.

The ObservableListComponentGeneratorEditorWindow has also been updated
so it sets the newly created GameObject name based on the component
selected rather than just the generic named `ComponentContainer` as
this makes it easier to identify newly created components.